- Entity statistic visibility modifier - moved other ent stat enums to correct package
#324
