+++ waltz-data/src/main/ddl/liquibase/db.changelog-1.0.xml
@@ -2214,5 +2214,23 @@
         </dropColumn>
     </changeSet>
 
+    <!-- 324 - Entit Stat Defn - Visibility Modifiers -->
+    <changeSet author="kamransaleem"
+               id="20160727-324-1">
+        <addColumn tableName="entity_statistic_definition">
+            <column name="entity_visibility"
+                    type="VARCHAR(32)"
+                    defaultValue="SHOW_ALL">
+                <constraints nullable="false"/>
+            </column>
+
+            <column name="rollup_visibility"
+                    type="VARCHAR(32)"
+                    defaultValue="SHOW_ALL">
+                <constraints nullable="false"/>
+            </column>
+        </addColumn>
+    </changeSet>
+
 
 </databaseChangeLog>
