+++ waltz-data/src/main/java/com/khartec/waltz/data/entity_statistic/EntityStatisticDefinitionDao.java
@@ -1,10 +1,10 @@
 package com.khartec.waltz.data.entity_statistic;
 
 import com.khartec.waltz.model.EntityKind;
-import com.khartec.waltz.model.StatisticCategory;
-import com.khartec.waltz.model.StatisticType;
 import com.khartec.waltz.model.entity_statistic.EntityStatisticDefinition;
 import com.khartec.waltz.model.entity_statistic.ImmutableEntityStatisticDefinition;
+import com.khartec.waltz.model.entity_statistic.StatisticCategory;
+import com.khartec.waltz.model.entity_statistic.StatisticType;
 import com.khartec.waltz.schema.tables.records.EntityStatisticDefinitionRecord;
 import org.jooq.*;
 import org.springframework.beans.factory.annotation.Autowired;
