+++ waltz-model/src/main/java/com/khartec/waltz/model/entity_statistic/StatisticVisibility.java
@@ -1,0 +1,7 @@
+package com.khartec.waltz.model.entity_statistic;
+
+public enum StatisticVisibility {
+    HIDE_ALL,
+    HIDE_EXEMPT,
+    SHOW_ALL
+}
